Revert "Handle POST /source/suite/{suiteId}/{serializedSuiteId} (#259)"
This reverts commit 76e7196517ca59de837ac34f5dab6bb5da603e77.
