+++ src/UrlGeneratorFactory.php
@@ -24,7 +24,6 @@ readonly class UrlGeneratorFactory
         $routeCollection->add('source', new Route('/source/{sourceId}', ['sourceId' => null]));
         $routeCollection->add('suite', new Route('/source/suite/{suiteId}', ['suiteId' => null]));
         $routeCollection->add('suites', new Route('/source/suites'));
-        $routeCollection->add('serialized-suite-create', new Route('/source/suite/{suiteId}/{serializedSuiteId}'));
 
         return new UrlGenerator($routeCollection, new RequestContext($baseUrl));
     }
