+++ tests/Integration/AbstractIntegrationTestCase.php
@@ -9,7 +9,6 @@ use GuzzleHttp\Psr7\HttpFactory;
 use PHPUnit\Framework\TestCase;
 use SmartAssert\ApiClient\Exception\Error\Factory as ExceptionFactory;
 use SmartAssert\ApiClient\Factory\Source\SourceFactory;
-use SmartAssert\ApiClient\Factory\Source\SuiteFactory;
 use SmartAssert\ApiClient\Factory\User\ApiKeyFactory;
 use SmartAssert\ApiClient\Factory\User\TokenFactory;
 use SmartAssert\ApiClient\Factory\User\UserFactory;
@@ -17,7 +16,6 @@ use SmartAssert\ApiClient\FileSourceClient;
 use SmartAssert\ApiClient\GitSourceClient;
 use SmartAssert\ApiClient\ServiceClient\HttpHandler;
 use SmartAssert\ApiClient\SourceClient;
-use SmartAssert\ApiClient\SuiteClient;
 use SmartAssert\ApiClient\UrlGeneratorFactory;
 use SmartAssert\ApiClient\UsersClient;
 use SmartAssert\ServiceRequest\Deserializer\Error\BadRequestErrorDeserializer;
@@ -40,7 +38,6 @@ abstract class AbstractIntegrationTestCase extends TestCase
     protected static FileSourceClient $fileSourceClient;
     protected static GitSourceClient $gitSourceClient;
     protected static SourceClient $sourceClient;
-    protected static SuiteClient $suiteClient;
 
     protected static ErrorDeserializer $errorDeserializer;
 
@@ -73,6 +70,5 @@ abstract class AbstractIntegrationTestCase extends TestCase
         self::$gitSourceClient = new GitSourceClient(new SourceFactory(), $httpHandler);
 
         self::$sourceClient = new SourceClient(new SourceFactory(), $httpHandler);
-        self::$suiteClient = new SuiteClient(new SuiteFactory(), $httpHandler);
     }
 }
